+++ lib/openstudio/analysis/translator/workflow.rb
@@ -2,9 +2,6 @@ module OpenStudio
   module Analysis
     module Translator
 
-      require 'json'
-      require 'securerandom'
-
       class Workflow
         attr_reader :osa_filename
         attr_reader :root_path
@@ -25,7 +22,7 @@ module OpenStudio
 
           # try to read the osa json file
           if File.exist?(@osa_filename)
-            @osa = ::JSON.parse(File.read(@osa_filename), {symbolize_names: true})[:analysis]
+            @osa = ::JSON.parse(File.read(@osa_filename), symbolize_names: true)[:analysis]
           else
             fail "File #{@osa_filename} does not exist"
           end
@@ -45,18 +42,23 @@ module OpenStudio
             step_hash = {}
             step_hash[:measure_dir_name] = File.basename(step[:measure_definition_directory])
             step_hash[:arguments] = {}
-            @osa[:problem][:workflow][i][:arguments].each do |arg|
-              step_hash[:arguments][arg[:name].to_sym] = arg[:value]
+            # Measures can have no arguments -- make sure to catch it
+            if @osa[:problem][:workflow][i][:arguments]
+              @osa[:problem][:workflow][i][:arguments].each do |arg|
+                step_hash[:arguments][arg[:name].to_sym] = arg[:value]
+              end
             end
             @steps << step_hash
           end
         end
 
+        # Convert a file in the form of an OSD into an OSW
         def process_datapoint(osd_filename)
           # Try to read the osd json file
+          osd = nil
           if File.exist?(osd_filename)
-            warn('data_point selector in ods will be changed to datapoint in version 1.0')
-            osd = ::JSON.parse(File.read(osd_filename), {symbolize_names: true})[:data_point]
+            # warn('data_point selector in ods will be changed to datapoint in version 1.0') # NL this isn't true anymore.
+            osd = ::JSON.parse(File.read(osd_filename), symbolize_names: true)[:data_point]
           else
             fail "File #{osd_filename} does not exist"
           end
@@ -78,7 +80,7 @@ module OpenStudio
           # Save the OSW hash
           osw = {}
           created_at = ::Time.now
-          osw[:seed_model] = @seed_file
+          osw[:seed_file] = @seed_file
           osw[:weather_file] = @weather_file
           osw[:file_format_version] = @osw_version
           osw[:osa_id] = @osa_id
@@ -93,13 +95,17 @@ module OpenStudio
 
         # Runs an array of OSD files
         def process_datapoints(osd_filename_array)
+          r = []
           osd_filename_array.each do |osd_file|
             begin
-              yield process_datapoint(osd_file)
+              r << process_datapoint(osd_file)
             rescue => e
-              puts "Warning: Failed to processes datapoint #{osd_file} with error #{e.message} in #{e.backtrace.join('\n')}"
+              r << nil
+              puts "Warning: Failed to process datapoint #{osd_file} with error #{e.message} in #{e.backtrace.join('\n')}"
             end
           end
+
+          return r
         end
       end
     end